Rename settings parser & validator.
